map_project/requirements.txt
asgiref==3.3.1
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
djangorestframework==3.12.2
folium==0.12.1
geographiclib==1.50
geopy==2.1.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.0
pandas==1.2.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
